Moved OperationDetail and OperationField to outer classes
